(react app) move components to component folder without subfolders
