+++ src/game_leader/robotwarsreact/src/components/NavigationBar.js
