+++ src/game_leader/robotwarsreact/src/components/RobotInfo.js
